+++ IzvorniKod/frontend/src/Components/snowfall/snowfall.css
@@ -1,0 +1,42 @@
+.snow-container {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 0;
+  height: 0;
+  pointer-events: none; /*pahuljice ne ometaju interakciju s ostalim elementima */
+  z-index: 10;
+}
+
+/*Snowflake styling*/
+.Snowflake {
+  position: absolute;
+  color: hsla(183, 70%, 34%, 0.879);
+  cursor: default; /*sprjecavam promjenu kursora kada prodem preko pahuljice*/
+  pointer-events: auto; /*omogucavam pointer-evente, da mogu maknuti pahuljicu*/
+  /*cubic-bezier kretanje nije linearno vec po krivulji (ubrzava pa usporava)*/
+  animation: fall cubic-bezier(.52, .09, .6, .97) infinite;
+  --rotateDirection: 1;
+}
+
+/*specifikacija animacije*/
+@keyframes fall {
+  0% {
+    opacity: 0;
+  }
+  3% {
+    opacity: 0.9;
+  }
+  90% {
+    opacity: 0.8;
+  }
+  100% {
+    transform: translateY(100vh) rotate(calc(360deg * var(--rotateDirection))); /* na dnu i zavrsena rotacija */
+    opacity: 0;
+  }
+}
+
+/*na hover unistimo objekt*/
+.Snowflake:hover {
+  display: none;
+}
