+++ IzvorniKod/frontend/src/main.jsx
@@ -1,8 +1,8 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
+import './styles/index.css'
 import App from './App.jsx'
-import {AuthProvider} from "./Firebase/AuthContext.jsx";
+import {AuthProvider} from "./firebase/AuthContext.jsx";
 
 createRoot(document.getElementById('root')).render(
 <AuthProvider>
