+++ IzvorniKod/frontend/src/styles/App.css
